backend/requirements.txt
django==5.1.1
gunicorn==23.0.0
mysqlclient==2.2.4
python-dotenv==1.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
